feat: move settings-provider to dapp folder
