+++ packages/dapp/src/components/index.ts
@@ -1,0 +1,6 @@
+export * from './BackgroundGradient';
+export * from './DualWidget';
+export * from './Navbar';
+export * from './SupportModal';
+export * from './ThemeSwitch';
+export * from './Widget';
