+++ packages/dapp/src/const/MenuConfig/MainMenu/Main-MenuItems.tsx
@@ -6,7 +6,6 @@ import TwitterIcon from '@mui/icons-material/Twitter';
 import { Typography } from '@mui/material';
 import { useTheme } from '@mui/material/styles';
 import { Discord, LifiSmallLogo } from '@transferto/shared/src/atoms/icons';
-import { useSettingsStore } from '@transferto/shared/src/contexts/SettingsContext';
 import { SettingsContextProps } from '@transferto/shared/src/types/settings';
 import { openInNewTab } from '@transferto/shared/src/utils/';
 import { useTranslation } from 'react-i18next';
@@ -14,6 +13,7 @@ import { SubMenuKeys } from '../../../const';
 import { useUserTracking } from '../../../hooks/useUserTracking/useUserTracking';
 import { useMenu } from '../../../providers/MenuProvider';
 import { useDetectDarkModePreference } from '../../../providers/ThemeProvider';
+import { useSettingsStore } from '../../../stores';
 
 export const useMainMenuItems = () => {
   const { t: translate, i18n } = useTranslation();
