+++ packages/dapp/src/stores/settings/index.ts
@@ -1,1 +1,2 @@
 export * from './SettingsContext';
+export * from './types';
