+++ packages/dapp/src/stores/settings/types.ts
@@ -1,0 +1,1 @@
+export const test = 'test';
