+++ packages/shared/src/index.ts
@@ -1,6 +1,5 @@
 export * from './atoms';
 export * from './config';
-export * from './contexts';
 export * from './hooks';
 // export * from './molecules';
 export * from './organisms';
